requirements.txt
streamlit==1.39.0
tensorflow==2.20.0
numpy
pillow
joblib
